korea_investment/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-apscheduler==0.6.2
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
